requirements.txt
anyio==4.3.0
asgiref==3.8.1
bcrypt==4.1.2
certifi==2024.2.2
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
dnspython==2.6.1
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_metadata==7.1.0
Markdown==3.6
ollama==0.1.8
pymongo==4.6.3
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1
zipp==3.18.1
